refacto Navbar
